Allow GitHub usernames as port maintainers
After the move to GitHub, it is sometimes difficult to figure out
whether a pull request was sent or a ticket was opened by a maintainer,
because both these systems use GitHub usernames now.

This leads to more confusion with developers and users than needed.
Change the documentation to allow GitHub usernames as maintainers for
Portfiles to avoid this.

Using a GitHub username leaves us with the problem that there is no
email address available for this user, unless the user has added
a public email address to their GitHub account. Email addresses are
instead available in commits, and maintainers can always be reached by
ticket or GitHub comment.