Add support for macro templates
This adds the possibility to define macros that can be reused across the
document. The first example for this is a macro that allows to place
a warning on sections that are outdated with this XML processing
instruction:

<?macro-outdated reason="..." ?>

This will be transformed using the template defined in
guide/xml/macros.xml, while the reason="" parameter passed to the
processing instruction will replace the placeholder written as
<macro-param name="reason" /> in the template.

This approach uses multiple passes of template evaluation and different
modes for the replacement. It is not possible to generate the macro
fragment within the XSL stylesheet itself, as in XSLT 1.0 there is
a huge difference between node sets from the document and fragments
generated on-the-fly. In order to allow <xref> inside the macro
template, the template itself has to be part of the document.