Describe GitHub changeset links in 7.1.3 Trac Ticket Guidelines
Describe the use of Trac changeset links, with a GitHub commit SHA
and repository name, as the preferred way to link to a commit in
the MacPorts repositories in GitHub.

There is a one-sentence summary in the TL;DR section, followed by
a one-paragraph explanation in the "Description" bullet item below.
There is a link to TracLinks#LinkstoMacPortsonGitHub in the wiki
for the full explanation.

Also, break the "Description" bullet item into two sub-bullets,
because there is a bit too much to read comfortably in one paragraph.

Change applied to the DocBook sources in guide/xml/project.xml, and
corresponding change applied to the adoc sources in
guide/adoc/project.adoc.