Modify reinplace term definiton to make optional files arg more clear
Before I saw it being used that way in someone else's Portfile, I had no idea that 'reinplace' is able to execute the regular expression on multiple files, because my eyes had completely ignored the "..." at the end of the term definition in the documentation. This commit makes it more clear that "..." means "additional files".