Remove experimental conversion to AsciiDoc
It has been several years since the last work to improve the yet
incomplete conversion to AsciiDoc. I originally wanted to have this in
the upstream repository to generate interest and involve more
contributors with the required manual fixes. However, I also lost the
focus on the guide conversion and did not continue the work.

At this point, having the unfinished conversion attempt in the
repository does more harm than it does good as it especially confuses
new contributors. Let's accept this is not happening and remove the
adoc/ directory.